Rename Issue/Redeem structs to match impl
This is a minor suggestion, aimed to bring the specs in closer alignment
with the implementation. The rationale being: it is easy to imagine a
design in which there there are two different structs, an `Issue` and an
`IssueRequest`, which represent different objects (e.g., if using an
event sourcing architecture). However, in the actual architecture
specified and implemented here, the `Issue` is just represented by an
`IssueRequest` with a certain status. (The same point holds mutatis
mutandis for `RedeemRequest`).

Since the collection of these requests is just named with the plural, it
suggests making singular requests is the most clarifying option.

Surfaced from @informalsystems' audit of interlay/btc-parachain at
75fd3e2